services/movies/requirements.txt
flask
flask_restful
webargs
flasgger
psycopg2-binary